Fix onFileSystem activation.
